test: isolate CLI docs privacy guard
